+++ examples/todo-pylist.html
@@ -28,23 +28,19 @@
         </nav>
         <section class="pyscript">
             <h1>To Do List</h1>
-            <py-tutor modules="utils.py;pylist.py">
-                <py-register-widget
-                    src="./pylist.py"
-                    name="py-list"
-                    klass="PyList"
-                ></py-register-widget>
-
+            <py-tutor modules="utils.py">
                 <py-config>
                     plugins = [
-                      "https://pyscript.net/latest/plugins/python/py_tutor.py"
+                      "https://pyscript.net/latest/plugins/python/py_tutor.py",
+                      "./py_list.py"
                     ]
                     [[fetch]]
-                    files = ["./utils.py", "./pylist.py"]
+                    files = ["./utils.py"]
                 </py-config>
 
                 <py-script>
                     from js import document
+                    from datetime import datetime as dt
                     from pyodide.ffi.wrappers import add_event_listener
 
                     def add_task(*args, **kws):
@@ -54,7 +50,8 @@
 
                       # add a new task to the list and tell it to use the `content` key to show in the UI
                       # and to use the key `done` to sync the task status with a checkbox element in the UI
-                      myList.add(task)
+                      myList = Element("myList")
+                      myList.element.pyElementInstance.add(task)
 
                       # clear the inputbox element used to create the new task
                       new_task_content.clear()
