+++ examples/todo.html
@@ -29,7 +29,8 @@
             <py-tutor modules="./utils.py;./todo.py">
                 <py-config>
                     plugins = [
-                      "https://pyscript.net/latest/plugins/python/py_tutor.py"
+                      "https://pyscript.net/latest/plugins/python/py_tutor.py",
+                      "./py_list.py"
                     ]
                     [[fetch]]
                     files = ["./utils.py", "./todo.py"]
