+++ pyscriptjs/src/python/pyscript/__init__.py
@@ -6,9 +6,6 @@ from ._event_loop import run_until_complete
 from ._html import (
     HTML,
     Element,
-    PyItemTemplate,
-    PyListTemplate,
-    PyWidgetTheme,
     add_classes,
     create,
     display,
@@ -43,9 +40,6 @@ __all__ = [
     "Element",
     "add_classes",
     "create",
-    "PyWidgetTheme",
-    "PyItemTemplate",
-    "PyListTemplate",
     "run_until_complete",
     "loop",
     "Plugin",
